Host login remake

+++ src/main/java/map/project/musiclibrary/service/HostUserService.java
@@ -1,7 +1,9 @@
 package map.project.musiclibrary.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import map.project.musiclibrary.data.model.audios.Podcast;
 import map.project.musiclibrary.data.model.users.HostUser;
+import map.project.musiclibrary.data.model.users.UserSession;
 import map.project.musiclibrary.data.repository.HostUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
@@ -24,27 +26,40 @@ public class HostUserService {
         this.podcastService = podcastService;
     }
 
-    public HostUser addHost(String name, String birthdateStr) throws ParseException {
-        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        Date birthdate = dateFormat.parse(birthdateStr);
+    public HostUser addHost(UserSession userSession, String name, String birthdateStr) {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+            Date birthdate;
+            try {
+                birthdate = dateFormat.parse(birthdateStr);
+            } catch (ParseException e) {
+                throw new IllegalArgumentException("Invalid birthdate format. Please use yyyy-MM-dd.");
+            }
 
-        HostUser host = (HostUser) CreatorUserFactory.createCreatorUser("host", name, birthdate);
+            HostUser host = (HostUser) CreatorUserFactory.createCreatorUser("host", name, birthdate);
 
-        return hostUserRepository.save(host);
+            return hostUserRepository.save(host);
+        }
+        throw new SecurityException("You must log in as an admin to add a host.");
     }
 
-    public void deleteHost(Long id) {
-        Optional<HostUser> hostUserOptional = hostUserRepository.findById(id);
+    public void deleteHost(UserSession userSession, String idStr) throws NumberFormatException {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            Long id = Long.parseLong(idStr);
+            Optional<HostUser> hostUserOptional = hostUserRepository.findById(id);
 
-        if (hostUserOptional.isPresent()) {
-            HostUser host = hostUserOptional.get();
+            if (hostUserOptional.isPresent()) {
+                HostUser host = hostUserOptional.get();
 
-            for (Podcast podcast : host.getPodcasts()) {
-                podcastService.deletePodcast(podcast.getId());
+                for (Podcast podcast : host.getPodcasts()) {
+                    podcastService.deletePodcast(podcast.getId());
+                }
+                host.getPodcasts().clear();
+                hostUserRepository.deleteById(id);
             }
-            host.getPodcasts().clear();
-            hostUserRepository.deleteById(id);
+            throw new EntityNotFoundException("Host with ID " + id + " was not found.");
         }
+        throw new SecurityException("You must log in as an admin to delete a host.");
     }
 
     public HostUser save(HostUser hostUser) {
@@ -55,8 +70,11 @@ public class HostUserService {
         return hostUserRepository.findByName(name).stream().findFirst().orElse(null);
     }
 
-    public List<HostUser> findAll() {
-        return hostUserRepository.findAll();
+    public List<HostUser> findAll(UserSession userSession) {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            return hostUserRepository.findAll();
+        }
+        throw new SecurityException("You must log in as an admin to list all hosts.");
     }
 
     public Optional<HostUser> findById(Long id) {
@@ -69,6 +87,6 @@ public class HostUserService {
         if (hostUserOptional.isPresent()) {
             return hostUserOptional.get().getPodcasts();
         }
-        throw new RuntimeException("HostUserService::Host with specified id doesn't exist");
+        throw new EntityNotFoundException("HostUserService::Host with specified id doesn't exist");
     }
 }
